using enum for SwiftUI view modifiers name; reorg
