+++ Stitch/Graph/StitchAI/Mapping/Syntax/RestrictedSwiftUIScopeJSON.swift
